task0-1

+++ 0x1A-hash_tables/hash_tables.h
@@ -1,5 +1,6 @@
 #ifndef HASH_TABLES
 #define HASH_TABLES
+
 /**
  * struct hash_node_s - Node of a hash table
  *
@@ -30,6 +31,6 @@ typedef struct hash_table_s
 } hash_table_t;
 
 
-hash_table_t *hash_table_create(unsigned int size);
+hash_table_t *hash_table_create(unsigned long int size);
 
 #endif
